+++ R/utils_define_constant.R
@@ -37,6 +37,16 @@ LASFIELDS             <- c("X", "Y", "Z", "Intensity",
                            "UserData", "gpstime", "PointSourceID",
                            "R", "G", "B")
 
+LASCATALOGATTRIBUTES <- c("File.Signature", "File.Source.ID", "GUID", "Version.Major",
+                          "Version.Minor", "System.Identifier", "Generating.Software",
+                          "File.Creation.Day.of.Year", "File.Creation.Year", "Header.Size",
+                          "Offset.to.point.data", "Number.of.variable.length.records",
+                          "Point.Data.Format.ID", "Point.Data.Record.Length", "Number.of.point.records",
+                          "X.scale.factor", "Y.scale.factor", "Z.scale.factor", "X.offset",
+                          "Y.offset", "Z.offset", "Max.X", "Min.X", "Max.Y", "Min.Y", "Max.Z",
+                          "Min.Z", "EPSG", "Number.of.1st.return", "Number.of.2nd.return",
+                          "Number.of.3rd.return", "Number.of.4th.return", "Number.of.5th.return")
+
 LIDRCONTEXTDSM <- "grid_canopy"
 LIDRCONTEXTSPI <- c("lasnormalize", "grid_terrain", "spatial_interpolation")
 LIDRCONTEXTGND <- "lasground"